Split off each directive with its node class into a separate file; create ItemElement parent class for custom node classes with utility functions
